pyproject.toml
[tool.poetry]
name = "aserehe"
version = "0.1.0"
description = "a semantic release helper CLI"
repository = "https://github.com/lukany/aserehe"
authors = ["Jan Lukány <lukany.jan@gmail.com>"]
license = "MIT"
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Framework :: Pytest",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: Utilities",
]

[tool.poetry.dependencies]
python = "^3.11"
GitPython = "^3.1.32"
semantic-version = "^2.10.0"
typer = "^0.9.0"

[tool.poetry.dev-dependencies]
black = "^23.7.0"
mypy = "^1.5.0"
pytest = "^7.4.0"
ruff = "^0.0.284"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
